Add CLI to the validator; return json object
