+++ actions/requirements.txt
@@ -2,3 +2,4 @@ beautifulsoup4==4.11.1
 validators==0.20.0
 lxml==4.9.1
 requests==2.28.1
+click
